implement api get statement scb

+++ Models/statement.py
@@ -3,5 +3,5 @@ from datetime import datetime
 class Statement():
     def __init__(self, date, amount, number):
         self.date: datetime = date
-        self.amount: int = amount
+        self.amount: float = amount
         self.number: str = number
